% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dhis2_secrets.R
\name{create_dhis2_secrets}
\alias{create_dhis2_secrets}
\title{Create DHIS2 Credentials as Keyring Secrets}
\usage{
create_dhis2_secrets()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Interactively prompts for a DHIS2 username and password and stores them
securely using the \pkg{keyring} package.
}
\details{
The secrets are saved under the keys \code{"dhis2_username"} and
\code{"dhis2_password"} in the default keyring.

This function is interactive and intended for local use only.
It should not be called in non-interactive sessions such as CI pipelines.
}
\examples{
\dontrun{
create_dhis2_secrets()
}

}
\seealso{
\code{\link[keyring]{key_set}}
}
